src/requirements.txt
appnope==0.1.4
asgiref==3.11.0
asttokens==3.0.0
attrs==24.3.0
certifi==2024.2.2
cffi==2.0.0
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
distlib==0.4.0
dj-database-url==3.0.1
Django==6.0
djangorestframework==3.16.1
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.20.0
fonttools==4.55.3
gunicorn==23.0.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
milvus==2.3.8
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
scipy==1.16.3
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
soundfile==0.13.1
sqlparse==0.5.4
srt==3.5.3
stack-data==0.6.3
stevedore==5.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.35.4
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
vosk==0.3.44
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.11.0
